fix: Bump vtk.js
vtk.js has changed what they export. They no longer export an object
called vtk, instead it is assigned to the window. So we have to change
our import and declare a variable vtk, that will be present at
runtime. This is not ideal. This can be cleaned up when we can get
rollup to pull in the parts of vtk we need.