Remove requested molecule from the redux state
In the mongochemclient, if a molecule page is opened by
clicking on a molecule card, it sees that the requested
molecule is in the current redux state and tries to
render it. However, the molecule does not have cjson until the
entire girder molecule is received, so it cannnot render
it.

This removes the requested molecule so that the first rendering
attempt is not performed.

This is especially important for determining if a molecule has
3D coordinates or not in the client, because a molecule is deemed
to not have 3D coordinates if any part of the cjson path to the 3D
coordinates is missing. This PR prevents the client from thinking
that a moleucle does not have 3D coordinates (and render an SVG)
when, in fact, it does.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>