Allow limiting of available API key permissions
Users can now pass a list (scopeOptions) to the API key module to select exacly
which scope options should be allowed to be set. Passing nothing in will default
to all available Girder options.

For example: if ['read', 'write'] is passed in only those permissions will be
displayed under the custom scope menu, and selecting 'Allow all' will set the
scope to only be 'read' and 'write'.

Signed-off-by: Brianna Major <brianna.major@kitware.com>